Translations (#2)
This is quite a big one, it was taken away from px_pharma_ui. Here's a summary of the changes:

* Initial commit

* PX-328 Add scripts for building the frontend

The idea is to customise it via an `assets` folder where we put all the stylesheets, images etc. that we need for a particular client.

There are two assumptions:

- you're compiling in the Vagrant VM;
- you've downloaded/cloned all the frontend repositories in the parent folder.

Steps to build:

    ./configure
    make
    make install

The final bundle will be in `/srv/px/CLIENT_NAME/ui` -- so in this case it's `/srv/px/px_pharma/ui`.

A couple of things to note:

- the Makefile is hardcoded for the test client for now; we could improve here by producing a customised Makefile for any client via `configure`.
- also the frontend modules are hardcoded in the Makefile, which is annoying because we're going to have to update it every time we create a new frontend module.

So it would definitely be an improvement to automate the Makefile creation. But it'd be overkill to do it now since we have only 2 frontend modules and one client.

* PX-328 Add `make build`

We want Jenkins to be as dumb as possible, so for building the frontend we'll just `git checkout && make build`.

* PX-328 Add assets folder with sample logos

* PX-559 Add help and default task to Makefile

* PX-559 We don't want to create symlinks in the package

* PX-559 Add packaging framework using FPM cookery

* PX-584 Add "make test" (#3)

* PX-603 Add "make checkout"

This command checks out a specific git ref for any frontend component, enabling us to manage multiple versions per customer.

* PX-603 Add "make clone"

* PX-559 Ensure assets are world readable for Nginx

* PX-559 Ensure post-install script is in the right location

* PX-584 Update Makefile (#7)

From now on we should be able to compile all the frontend components by running `make build`. Same if we want to compile just one single component.

Related to:

- https://github.com/MyMedsAndMe/px_data_capture_ui/pull/413
- https://github.com/MyMedsAndMe/px_authorisation_ui/pull/73

* PX-559 Ensure permissions are correct with post-install script (#9)

* PX-621 Remove logos (#10)

* PX-621 Add assets folder (#11)

* PX-540 Add translations elixir app

* PX-540 Add translations source

* PX-540 Fix template spacing

* PX-612 Build the translations binary and generate translations

* PX-540 git rm --cached translations/translations

* PX-540 wording

* PX-540 Add .gitignore

* PX-540 Downgrade Elixir

Because the version we use in the Vagrant VM is 1.2.6, and mix was complaining that:

** (Mix) You're trying to run :translations on Elixir v1.2.6 but it has declared in its mix.exs file it supports only Elixir ~> 1.3

* PX-612 Delete priv folder

It was there just for testing.

* PX-540 Get path from argument

* PX-612 Configurable path to gettext

* PX-612 Provide path to gettext folder

* PX-612 Use only languages selected by user (incomplete)

* PX-612 Disable authorisation_ui translations

This is a temporary fix.

* PX-540 Remove extra line and fix spacing

* PX-540 Put last language on top

I don't know why this is happening, but I had to do this because otherwise the translations would stop working. The thing is, the current implementation is order-dependent, so if we don't get the languages in alphabetical order it won't work.

* PX-540 Display "Missing translation key" if the key is not empty

* PX-612 Generalising code

* PX-612 Update template

* PX-612 Refactoring

* PX-612 Adapt locales

* PX-612 Update .gitignore

* PX-612 Refactoring

* PX-612 Add unit tests

* PX-612 Refactoring tests

* PX-612 Fix import Dict

* PX-612 delete unrelated files from px_pharma_ui

* PX-612 move to parent folder

* PX-612 Update README.md

* PX-612 Work with an absolute path